requirements.txt
anthropic==0.40.0
google-genai>=0.1.0
python-dotenv==1.0.1
PyPDF2==3.0.1
streamlit==1.32.0
notion-client==2.2.1 